Separate PlayerService
It now exists as PlayerCreateService and PlayerGetService.
