+++ ShapeDungeon/Interfaces/Services/Players/IPlayerGetService.cs
@@ -1,0 +1,10 @@
+﻿using ShapeDungeon.DTOs.Players;
+
+namespace ShapeDungeon.Interfaces.Services.Players
+{
+    public interface IPlayerGetService
+    {
+        Task<IEnumerable<PlayerDto>> GetAllPlayersAsync();
+        Task<PlayerDto> GetPlayerAsync(string name);
+    }
+}
